Linspace operation

+++ lib/mlirlowerer/test/src/InputOutputTest.cpp
@@ -549,7 +549,7 @@ TEST(Output, integerArrayWithSizeDependingOnInputValue)	 // NOLINT
 	jit::Runner runner(*module);
 	ASSERT_TRUE(mlir::succeeded(runner.run("main", x, jit::Runner::result(yPtr))));
 
-	for (int i = 0; i < yPtr.getDimensionSize(0); i++)
+	for (size_t i = 0; i < yPtr.getDimensionSize(0); i++)
 		EXPECT_EQ(yPtr[i], i + 1);
 }
